@@ -20,9 +20,10 @@
         }
     },
     "require-dev": {
-        "phpunit/phpunit": "^6.4"
+        "phpunit/phpunit": "^9.0"
     },
     "require": {
-        "php": "^7.1"
+        "php": "^8.0",
+        "ext-iconv": "*"
     }
 }
